상품 소개 디테일 컴포넌트화
