+++ lib/screens/01_home/HomeScreen.dart
@@ -1,8 +1,7 @@
 import 'package:daangn/models/product.dart';
-import 'package:daangn/theme.dart';
+import 'package:daangn/screens/01_home/components/product_detail.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({
@@ -78,56 +77,9 @@ class PrductItem extends StatelessWidget {
           ),
           SizedBox(width: 15.0),
           /* 상품 티테일 */
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(product.title, style: textTheme().bodyLarge),
-              SizedBox(height: 5.0),
-              Text("${product.address} · ${product.publishedAt}"),
-              SizedBox(height: 5.0),
-              Text("${numberFormat(product.price)}원"),
-              Spacer(),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.end /* TODO : 뭐지 */,
-                children: [
-                  Visibility(
-                    visible: product.commentsCount > 0 /* 0 이상 이어야 말풍선 표시 */,
-                    child: _buildIcons(
-                      product.commentsCount,
-                      CupertinoIcons.chat_bubble_2,
-                    ),
-                  ),
-                  const SizedBox(width: 8.0),
-                  Visibility(
-                    visible: product.heartCount > 0,
-                    child: _buildIcons(
-                      product.heartCount,
-                      CupertinoIcons.heart,
-                    ),
-                  ),
-                ],
-              )
-            ],
-          )
+          ProductDetail(product: product)
         ],
       ),
-    );
-  }
-
-  /* NumberFormat 임포트 x */
-  String numberFormat(String price) {
-    final formatter = NumberFormat('#,###');
-    return formatter.format(int.parse(price));
-  }
-
-  /* 아이콘과 숫자 표시*/
-  Widget _buildIcons(int count, IconData iconData) {
-    return Row(
-      children: [
-        Icon(iconData, size: 14.0),
-        const SizedBox(width: 4.0),
-        Text('$count'),
-      ],
     );
   }
 }
